# app/web/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "Cultural Screenplay & Visual Adaptation Studio",
  "version": "0.1.0"
 },
 "paths": {
...
  "/api/projects": {
   "get": {
    "summary": "List Projects",
    "operationId": "list_projects_api_projects_get",
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "items": {
          "$ref": "#/components/schemas/ProjectSummary"
         },
         "type": "array",
         "title": "Response List Projects Api Projects Get"
        }
       }
      }
     }
    }
   },
   "post": {
    "summary": "Create Project",
    "operationId": "create_project_api_projects_post",
    "requestBody": {
     "content": {
      "multipart/form-data": {
       "schema": {
        "$ref": "#/components/schemas/Body_create_project_api_projects_post"
       }
      }
     },
     "required": true
    },
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Created"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/projects/{pid}": {
   "get": {
    "summary": "Get Project",
    "operationId": "get_project_api_projects__pid__get",
    "parameters": [
     {
      "name": "pid",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Pid"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ProjectView"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/projects/{pid}/job": {
   "get": {
    "summary": "Get Job",
    "operationId": "get_job_api_projects__pid__job_get",
    "parameters": [
     {
      "name": "pid",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Pid"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "anyOf": [
          {
           "$ref": "#/components/schemas/JobInfo"
          },
          {
           "type": "object",
           "additionalProperties": true
          }
         ],
         "title": "Response Get Job Api Projects  Pid  Job Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/projects/{pid}/run/{stage}": {
   "post": {
    "summary": "Run Stage",
    "operationId": "run_stage_api_projects__pid__run__stage__post",
    "parameters": [
     {
      "name": "pid",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Pid"
      }
     },
     {
      "name": "stage",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Stage"
      }
     },
     {
      "name": "force",
      "in": "query",
      "required": false,
      "schema": {
       "type": "boolean",
       "default": false,
       "title": "Force"
      }
     }
    ],
    "responses": {
     "202": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/JobInfo"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/projects/{pid}/replan/scene/{scene_id}": {
   "post": {
    "summary": "Replan One Scene",
    "operationId": "replan_one_scene_api_projects__pid__replan_scene__scene_id__post",
    "parameters": [
     {
      "name": "pid",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Pid"
      }
     },
     {
      "name": "scene_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Scene Id"
      }
     }
    ],
    "responses": {
     "202": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/JobInfo"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/projects/{pid}/rewrite/scene/{scene_id}": {
   "post": {
    "summary": "Rewrite One Scene",
    "description": "Regenerate one scene's rewrite, bypassing the cache (an explicit \"try again\").",
    "operationId": "rewrite_one_scene_api_projects__pid__rewrite_scene__scene_id__post",
    "parameters": [
     {
      "name": "pid",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Pid"
      }
     },
     {
      "name": "scene_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Scene Id"
      }
     }
    ],
    "responses": {
     "202": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/JobInfo"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/projects/{pid}/screenplay": {
   "get": {
    "summary": "Screenplay",
    "operationId": "screenplay_api_projects__pid__screenplay_get",
    "parameters": [
     {
      "name": "pid",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Pid"
      }
     },
     {
      "name": "gloss",
      "in": "query",
      "required": false,
      "schema": {
       "type": "boolean",
       "default": false,
       "title": "Gloss"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "text/plain": {
        "schema": {
         "type": "string"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/projects/{pid}/comparison": {
   "get": {
    "summary": "Comparison",
    "description": "Source next to adapted, per scene, with the decisions that explain the changes.",
    "operationId": "comparison_api_projects__pid__comparison_get",
    "parameters": [
     {
      "name": "pid",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Pid"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "array",
         "items": {
          "$ref": "#/components/schemas/ComparisonScene"
         },
         "title": "Response Comparison Api Projects  Pid  Comparison Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/projects/{pid}/replan/entities": {
   "post": {
    "summary": "Replan All Entities",
    "operationId": "replan_all_entities_api_projects__pid__replan_entities_post",
    "parameters": [
     {
      "name": "pid",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Pid"
      }
     }
    ],
    "responses": {
     "202": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/JobInfo"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/projects/{pid}/recheck": {
   "post": {
    "summary": "Recheck",
    "operationId": "recheck_api_projects__pid__recheck_post",
    "parameters": [
     {
      "name": "pid",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Pid"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ProjectView"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/projects/{pid}/characters/{cid}": {
   "patch": {
    "summary": "Edit Character",
    "operationId": "edit_character_api_projects__pid__characters__cid__patch",
    "parameters": [
     {
      "name": "pid",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Pid"
      }
     },
     {
      "name": "cid",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Cid"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "type": "object",
        "additionalProperties": true,
        "title": "Fields"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/AffectedScenes"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/projects/{pid}/merge": {
   "post": {
    "summary": "Merge",
    "operationId": "merge_api_projects__pid__merge_post",
    "parameters": [
     {
      "name": "pid",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Pid"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/MergeBody"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/AffectedScenes"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/projects/{pid}/plan/{kind}/{eid}": {
   "patch": {
    "summary": "Edit Plan",
    "operationId": "edit_plan_api_projects__pid__plan__kind___eid__patch",
    "parameters": [
     {
      "name": "pid",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Pid"
      }
     },
     {
      "name": "kind",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Kind"
      }
     },
     {
      "name": "eid",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Eid"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "type": "object",
        "additionalProperties": true,
        "title": "Fields"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/AffectedScenes"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/projects/{pid}/plan/scenes/{scene_id}/keep": {
   "post": {
    "summary": "Keep Scene",
    "operationId": "keep_scene_api_projects__pid__plan_scenes__scene_id__keep_post",
    "parameters": [
     {
      "name": "pid",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Pid"
      }
     },
     {
      "name": "scene_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Scene Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/projects/{pid}/decisions/{did}/review": {
   "post": {
    "summary": "Review Decision",
    "operationId": "review_decision_api_projects__pid__decisions__did__review_post",
    "parameters": [
     {
      "name": "pid",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Pid"
      }
     },
     {
      "name": "did",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Did"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/ReviewBody"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/AdaptationDecision"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
  "/api/projects/{pid}/blockers/{what}": {
   "get": {
    "summary": "Get Blockers",
    "operationId": "get_blockers_api_projects__pid__blockers__what__get",
    "parameters": [
     {
      "name": "pid",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Pid"
      }
     },
     {
      "name": "what",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "What"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Blockers"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
...
   "ProjectSummary": {
    "properties": {
     "id": {
      "type": "string",
      "title": "Id"
     },
     "source_filename": {
      "type": "string",
      "title": "Source Filename"
     },
     "status": {
      "type": "string",
      "title": "Status"
     },
     "created_at": {
      "type": "string",
      "title": "Created At"
     },
     "culture": {
      "type": "string",
      "title": "Culture"
     },
     "region": {
      "type": "string",
      "title": "Region"
     },
     "scenes": {
      "type": "integer",
      "title": "Scenes"
     },
     "characters": {
      "type": "integer",
      "title": "Characters"
     },
     "decisions": {
      "type": "integer",
      "title": "Decisions"
     },
     "flagged_decisions": {
      "type": "integer",
      "title": "Flagged Decisions"
     },
     "unreviewed_decisions": {
      "type": "integer",
      "title": "Unreviewed Decisions",
      "default": 0
     },
     "job": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/JobInfo"
       },
       {
        "type": "null"
       }
      ]
     }
    },
    "type": "object",
    "required": [
     "id",
     "source_filename",
     "status",
     "created_at",
     "culture",
     "region",
     "scenes",
     "characters",
     "decisions",
     "flagged_decisions"
    ],
    "title": "ProjectSummary"
   },
...
  }
 }
}